looks like the only cleanups I know are valgrinding and increasing warnings

+++ source/Codegen/ArrayCodegen.cpp
@@ -40,7 +40,7 @@ Result_t ArrayIndex::codegen(Codegen::CodegenInstance* cgi)
 			if((n = dynamic_cast<Number*>(this->index)))
 			{
 				assert(!n->decimal);
-				if(n->ival >= at->getNumElements())
+				if((uint64_t) n->ival >= at->getNumElements())
 					error(this, "Compile-time bounds checking detected index '%d' is out of bounds of %s[%d]", n->ival, this->var->name.c_str(), at->getNumElements());
 			}
 		}
